\name{baselineCorrectionQuant}
\alias{baselineCorrectionQuant}
\title{Baseline correction - quantiles method}
\description{
    This function perform baseline correction using a quantiles around a
    moving window algorthim.
}
\usage{
    baselineCorrectionQuant(peaks, time, smooth=0, qntl=0.50, width=30,
                            unit=c("seconds", "points"), steps=10)
}
\arguments{
  \item{peaks}{Either a matrix object of spectra peak intensities to be baseline corrected,
    where the rows are retention times and columns are mass traces; or, a named list containing
    an element called \code{"Peaks"} which such matrix and another called \code{"Time"} with
    the retention time in seconds. The list can be generated by \code{\link{peakCDFextraction}}}
  \item{time}{A vector of retention time in seconds. This parameter is used if \code{peaks}
    is a matrix. Otherwise, the element called \code{"Time"} is used instead and this parameter
    is ignored.}
  \item{smooth}{An integer. Smooth each signal by this number of points using a moving average.
    Smoothing is disabled if this value is less or equal than 1. Note that the smoothing is applied
    after the baseline correction.}
  \item{qntl}{Numeric scalar. The quantile for baseline estimation. The value must be in [0, 1].}
  \item{width}{numeric scalar. The size of the window centered around a scan for baseline estimation.
    The size depends on the parameter \code{unit} below.}
  \item{unit}{A string which chooses if the \code{width} are points (scans) or seconds.}
  \item{steps}{Integer scalar greater than zero. To speed up computation, the baseline
   algorithm does not compute the baseline estimate in each single scan, but in intervals of
   \code{steps}  steps. The intermediate points are estimated by simple linear regression.}
}
\details{
  Applies a quantile based baseline estimation method. The method is applied for each ion
  mass trace (column of peaks) individually. It simple computes for each data point of the
  trace the \code{qntl} quantile, for example the 50\% quantile, ie, the median, of all
  the points which are within a \code{width} distance or it.

  In order for the method to work, select a \code{width} much larger than the widest peak.

  For speed efficiency, and assuming that the baseline is a smooth curve, the quantiles
  are computed every \code{step} points instead. By setting this value to 1, it means
  compute the baseline for each point. The baseline of the points in between are approximated
  by linear interpolation.
}
\value{
  Returns a list with the same elements as the the input, but the element \code{"Peaks"}
  containing baseline corrected values. In case peaks is a matrix, it returns a matrix
  of the same dimension instead.
}
\author{Alvaro Cuadros-Inostroza}
\seealso{ \code{\link{RIcorrect}}, \code{\link{baseline}}, \code{\link{baselineCorrection}} }
