\name{TargetSearch-defunct}
\alias{TargetSearch-defunct}

\alias{fixRIcorrection}
\alias{fixRIcorrection-defunct}

\title{Defunct functions in package \pkg{TargetSearch}}
\description{
    Functions listed here have been made defunct.
}

\usage{
    fixRIcorrection(...)
}
\details{
    \code{fixRIcorrection} has been replaced by \code{\link{fixRI}}.
}
