\name{TargetSearch-package}
\alias{TargetSearch-package}
\alias{TargetSearch}
\docType{package}
\title{
A targeted approach for GC-MS data.
}
\description{
This packages provides a targeted method for GC-MS data analysis. The workflow
includes a peak picking algorithm to convert from netcdf files to tab delimited
files, retention time correction using retention time markers provided by the
user, and a library search using multiple marker masses and retention time index
optimisation.
}
\author{
Alvaro Cuadros-Inostroza,
Matthew Hannah,
Henning Redestig

Maintainer: Alvaro Cuadros-Inostroza <inostroza@mpimp-golm.mpg.de>
}

