% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRimLim.R
\name{checkRimLim}
\alias{checkRimLim}
\title{Visually check retention index marker limits}
\usage{
checkRimLim(samples, rim, layout, show = TRUE, extend = 0.5,
  rect.col = "#e7e7e7", mar = c(2, 2, 2, 2), oma = c(3, 3, 2, 0.5),
  cex.main = 1, type = "l", ...)
}
\arguments{
\item{samples}{A  \code{\linkS4class{tsSample}} object created by
\code{\link{ImportSamples}}.}

\item{rim}{A \code{\linkS4class{tsRim}} object describing the retention index
markers. See \code{\link{ImportFameSettings}}.}

\item{layout}{A vector of the form \code{c(nr, nc)} to arrange the panel by
\code{nr} rows and \code{nc} columns. If missing then the layout is created
automatically.}

\item{show}{Logical. If \code{FALSE} the plot is not shown, but the data points
can be used for further inspection or for custom plots.}

\item{extend}{a numeric coefficient to extend the time window search of the
respective time marker. Defaults to \code{0.5}.}

\item{rect.col}{the color for the background rectangle which indicates
the current retention time limits.}

\item{mar}{the subplots margins, passed to \code{\link[graphics]{par}()}.}

\item{oma}{the outer plot margins, passed to \code{\link[graphics]{par}()}.}

\item{cex.main}{The magnification to be used for main titles, passed to
\code{\link[graphics]{par}()}.}

\item{type}{A character vector indicating the type of plots. Default \code{"l"} for
lines. Passed to \code{\link[graphics]{plot}()}.}

\item{...}{extra plotting arguments passed to \code{\link[graphics]{plot}()}
such as \code{col}, \code{lty}, \code{pch}, \code{lwd}.}
}
\value{
A list of \code{n} times \code{2} matrices or invisible. Each element correspond to a
   marker. Columns are retention time and intensities of the respective marker's m/z.
   The rows can be as many data points are within the search window.
}
\description{
A function to visually check if the retention time limites of the retention
index markers (aka FAMEs) is correct.
}
\details{
The function takes a random CDF file from your \code{\linkS4class{tsSample}}
object and creates a panel plot of the m/z traces around the area in which a
marker is expected to be. Repeated calls to this function can be used to check
other samples. It is also possible to check a specific sample by indexing
the \code{\linkS4class{tsSample}} object.
}
\examples{
require(TargetSearchData)

# get the cdf path TargetSearchData
cdfpath <- file.path(find.package("TargetSearchData"), "gc-ms-data")

# import samples (see ImportSamples() for details)
samples <- ImportSamples(file.path(cdfpath, "samples.txt"), CDFpath = cdfpath)

# Import RI markers (see ImportFameSettings())
rim <- ImportFameSettings(file.path(cdfpath, "rimLimits.txt"))

# choose a sample at random and plot the m/z traces around the retention
# time window
ret <- checkRimLim(samples, rim)

}
\seealso{
\code{\linkS4class{tsSample}}, \code{\linkS4class{tsRim}}, \code{\link{ImportFameSettings}}
}
