% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF.R
\name{ncdf4_convert_from_path}
\alias{ncdf4_convert_from_path}
\title{Convert CDF files to CDF4 from a path automatically}
\usage{
ncdf4_convert_from_path(cdf_path = ".", out_path = cdf_path, ...)
}
\arguments{
\item{cdf_path}{the input path to scan for}

\item{out_path}{the output path in which the files will be saved}

\item{\dots}{extra options passed to \code{\link[=ncdf4_convert]{ncdf4_convert()}}, which in
can be passed to \code{\link[=baseline]{baseline()}}}
}
\value{
a character vector of generated files or invisible.
}
\description{
Convert from NetCDF format 3 into a custom TargetSearch NetCDF format 4
automatically by scanning for CDF-3 files in given path and calling
the function \code{\link[=ncdf4_convert]{ncdf4_convert()}} on them.
}
\details{
This function simply wraps around \code{\link[=ncdf4_convert]{ncdf4_convert()}}. It searches for
CDF-3 files and converts them to CDF-4. Note that the search is not
recursive.

By default, the function scans the current path and saves them in
the same path if \code{out_path} is not specified.
}
\examples{
\dontrun{
# get files from package TargetSearchData
require(TargetSearchData)
cdfpath <- file.path(find.package("TargetSearchData"), "gc-ms-data")
ncdf4_convert_from_path(cdfpath, ".")
}

}
\seealso{
\code{\link[=ncdf4_convert]{ncdf4_convert()}}, \code{\link[=baseline]{baseline()}}
}
\author{
Alvaro Cuadros-Inostroza
}
