% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF.R
\name{ncdf4_update_ri}
\alias{ncdf4_update_ri}
\title{Update retention time index on a NCDF4 file}
\usage{
ncdf4_update_ri(cdfFile, observed, standard)
}
\arguments{
\item{cdfFile}{Path to the CDF file}

\item{observed}{The observed RI markers retention times'.}

\item{standard}{The RI of said markers.}
}
\value{
Returns \code{invisible}
}
\description{
Performs retention time index (RI) correction on a CDF file, using the
retention markers found by \code{\link[=RIcorrect]{RIcorrect()}}, or to force the markers time
if, for example, the RI markers were not co-injected with the biological
samples. It wraps around \code{\link[=rt2ri]{rt2ri()}}
}
\details{
This function is similar to \code{\link[=fixRI]{fixRI()}}, with the difference that is acts
upon a single file, whereas \code{\link[=fixRI]{fixRI()}} requires a \linkS4class{tsSample}
object.
}
\note{
This function is meant to be used internally. It is exposed for convenience.
}
\seealso{
\code{\link[=fixRI]{fixRI()}}, \code{\link[=RIcorrect]{RIcorrect()}}
}
\author{
Alvaro Cuadros-Inostroza
}
