\name{TargetSearch-defunct}
\alias{TargetSearch-defunct}

\alias{fixRIcorrection}
\alias{fixRIcorrection-defunct}
\alias{TargetSearchGUI}

\title{Defunct functions in package \sQuote{TargetSearch}}

\description{
  Functions listed here are defunct and no longer available.
}
\details{
  These function have been removed from \sQuote{TargetSearch} and no
  longer available. Use the replacement if any.

  \itemize{
    \item \code{fixRIcorrection} has been replaced by \code{\link{fixRI}}.

    \item \code{TargetSearchGUI} has been removed. Please use regular
    \sQuote{TargetSearch} functions. The old source code is archived at
    \url{https://www.github.com/acinostroza/TargetSearchGUI}.
  }
}
