\name{TargetSearch-deprecated}
\alias{TargetSearch-deprecated}

\title{Deprecated functions in package \sQuote{TargetSearch}}

\description{
    The functions listed here have been deprecated and may be
    removed as soon as of the next release.
}
\details{
    These functions are deprecated. Use the replacement if available.
    \itemize{
    \item \code{\link{libId}} method has been deprecated, or rather
    renamed. Use the fully equivalent \code{\link{makeIndex}}.
    }
}
