\name{Write.Results}
\alias{Write.Results}
\title{Save TargetSearch result objects into files}
\description{
  This is a convenient function to save the TargetSearch result into text files.
}
\usage{
Write.Results(Lib, metabProfile, prefix = NA)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Lib}{ A \code{tsLib} object. }
  \item{metabProfile}{ A \code{tsProfile} object. The final result of the package. This object is generated by
		either \code{Profile} or \code{ProfileCleanUp}. }
  \item{prefix}{ A character string. This is used as a name prefix for the written files.
	\code{"TargetSearch-"} is used by default. }
}
\value{
	This function doesn't return anything. Just print a message with the saved files.
}
\author{Alvaro Cuadros-Inostroza, Matthew Hannah, Henning Redestig }
\seealso{ \code{\link{peakFind}}, \code{\link{Profile}}, \code{\link{ProfileCleanUp}},
\code{\linkS4class{tsLib}}, \code{\linkS4class{tsMSdata}}, \code{\linkS4class{tsProfile}} }
